requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pyyaml==6.0.1
pydantic==2.7.1
pytest==8.2.0
httpx==0.27.0
google-generativeai==0.5.4
python-multipart==0.0.9
python-dotenv